Split up generate image renditions to run simulaneously
